ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr==1.7.0
openai==1.3.0
pillow>=10.1.0
python-dotenv==1.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
